requirements.txt
# LangGraph & Core AI
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0

# Model Context Protocol
mcp>=1.4.0

# Vector Database (RAG)
chromadb>=0.5.0

# Environment Management
python-dotenv>=1.0.0

# Web Server & UI Streaming
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.1.0